Bayes Factor Design Analysis Precalculation Results

Dataset that contains precalculated sample sizes with the BFDA method. We ran each calculation with 10000 iteration. The dataset contains the input and output values of these calculations.

bfda_precalculation_results

Format

A dataframe with 1230 rows and 9 variables:

iterate

numeric, unique id of iteration

tpr

numeric, The long-run probability of obtaining a Bayes factor at least as high as the critical threshold favoring superiority, given Delta.

delta

numeric, The expected population effect size.

thresh

numeric, The Bayes factor threshold for inference. Either 3, 6 or 10.

n1

numeric, The determined sample size per group.

tpr_out

numeric, The TPR associated with the resulting sample sizes.

h0

numeric, The frequency of the Bayes factor providing evidence for the null hypothesis with the given threshold.

ha

numeric, The frequency of the Bayes factor providing evidence for the alternative hypothesis with the given threshold.

error_message

character, The error message in case of an error.

Remark

If none of the precalculated values suit your sample size determination plan, than feel free the calculations with the R package by using the ssp_bfda function.