Equivalence Interval Bayes Factor Precalculation Results

Dataset that contains precalculated sample sizes with the equivalence interval Bayes factor method. The dataset contains the input and output values of these calculations.

Usage

eq_bf_precalculation_results

Format

A dataframe with 33210 rows and 9 variables:

iterate

numeric, unique id of iteration

tpr

numeric, The long-run probability of obtaining a Bayes factor at least as high as the critical threshold favoring superiority, given Delta.

delta

numeric, The expected population effect size.

thresh

numeric, The Bayes factor threshold for inference. Either 3, 6, or 10.

prior_scale

numeric, Scale of the Cauchy prior distribution. Either 1/sqrt(2), 1, or sqrt(2).

eq_band

numeric, The chosen width of the equivalence region.

n1

numeric, The determined sample size per group.

tpr_out

numeric, The TPR associated with the resulting sample sizes.

error_message

character, The error message in case of an error.

Remark

If none of the precalculated values suit your sample size determination plan, than feel free the calculations with the R package by using the ssp_eq_bf function.