Determine sample size with Accuracy In Parameter Estimation (AIPE) method

A sample size estimation method used for parameter estimation. The approach aims to find the required sample size, such that the confidence interval has a certain expected width. The function is a simple wrapper function around the ss.aipe.smd function.

ssp_aipe(
  delta,
  confidence_level,
  width,
  which_width = "Full",
  certainty = NULL,
  ...
)

Arguments

delta
Numeric. The expected population effect size.

width
Numeric. The desired width of the confidence interval, given Delta.

which_width
Character. The desired width of interest. Either `Full`, `Lower` or `Upper`.

...
Other arguments to be passed to the ss.aipe.smd function.

confidence_levell
Numeric. The desired level of confidence.

certainity
Numeric. The desired certainty of the confidence interval width.

Value

The function returns a list of one named element. The determined sample size for group one and two called `n1`.

Examples

if (FALSE) {
SampleSizePlanner::ssp_aipe(delta = 0.5, width = 0.2, confidence_level = 0.8)
}